MemStruct: auto-allocation feature
MemStruct can be automatically allocated if a None addr is passed to the
constructor and mem.allocator has been set to an allocation function.

miasm2.os_dep.common.heap API has been extended to directly support a
VmMngr as an argument.

NOTE: heap.alloc and heap.vm_alloc could be merged, but allowing the
first argument to be either a jitter or a vm is misleading, and changing
the old API would have broken some code.